[SP-3631] inject ConnectivityManager on MessageWebVC
